refactor: CommentInputForm에서 직접 mutation 한다
- CommentPage에서 mutaion을 하지 않고 CommentInputForm이 직접 mutation을 하도록한다.
- relay의 co-location 을 활용한다
